Modification of the way we define y_soft
y_soft used to be [0, ..., 0, 0.65, 0, ..., 0], with 0.65 at the indice of the hard label. So taking the argmax of that would simply give the hard label and not soft label. What we did,is that we computed the dot product between y and each embedded class label, and apply a soft max, so that we can have y_soft like [0.2, 0.4, 0.5, 0.1]. Taking the argmax would give you the soft label